@@ -2,27 +2,27 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.4",
+        "php": "^8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.5",
-        "doctrine/orm": "^2.11",
-        "easycorp/easyadmin-bundle": "^2.1",
+        "doctrine/annotations": "^1.14",
+        "doctrine/doctrine-bundle": "^2.10",
+        "doctrine/orm": "^2.16",
+        "easycorp/easyadmin-bundle": "^2.3",
         "mailjet/mailjet-apiv3-php": "^1.5",
-        "ovh/ovh": "^2.0",
+        "ovh/ovh": "^2.1",
         "sensio/framework-extra-bundle": "^6.2",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "5.4.*",
         "symfony/console": "5.4.*",
         "symfony/dotenv": "5.4.*",
         "symfony/expression-language": "5.4.*",
-        "symfony/flex": "^1.1",
+        "symfony/flex": "^1.20",
         "symfony/form": "5.4.*",
         "symfony/framework-bundle": "5.4.*",
         "symfony/http-client": "5.4.*",
         "symfony/mailer": "5.4.*",
-        "symfony/monolog-bundle": "^3.7",
+        "symfony/monolog-bundle": "^3.8",
         "symfony/process": "5.4.*",
         "symfony/property-access": "5.4.*",
         "symfony/property-info": "5.4.*",
@@ -35,12 +35,12 @@
         "symfony/validator": "5.4.*",
         "symfony/web-link": "5.4.*",
         "symfony/yaml": "5.4.*",
-        "vich/uploader-bundle": "^1.19"
+        "vich/uploader-bundle": "^1.23"
     },
     "require-dev": {
         "rector/rector": "^0.18.0",
         "symfony/debug-bundle": "5.4.*",
-        "symfony/maker-bundle": "^1.11",
+        "symfony/maker-bundle": "^1.50",
         "symfony/stopwatch": "5.4.*",
         "symfony/web-profiler-bundle": "5.4.*"
     },
